better place for test case
